Foe item details component, add a rolling button
